requirements.txt
torch>=2.1.0
torchvision>=0.16
timm>=0.9.12
pycocotools>=2.0.7
pyyaml
tqdm
tensorboard
albumentations>=1.4.1
opencv-python-headless
